Refactor testing to add more exporters
